Split linearisation and hierarchy checking
Results in simpler and faster code.
